# sdk/openapi.json
{
  "components": {
    "schemas": {
...
      "ChainStatus": {
        "description": "Mirrors lib/chain-status.ts's ChainStatus (LEGACY_OFFCHAIN +\nChainStatusBucket) value-for-value \u2014 same reason StatusKey above\nmirrors types.ts: that file already collapses GenLayer's real 14-value\nTransactionStatus into this 4-bucket + legacy shape (via the SDK's own\nisDecidedState(), not a hand-copied list \u2014 see that file's header),\nand duplicating a *different* vocabulary here would let the two\ndrift. scripts/genlayer-read.ts computes the bucket (it has\ngenlayer-js loaded); services/genlayer_indexer.py only ever writes\none of these five strings into a row's chain_status column.\n\nEvery Milestone/Dispute/Prediction row defaults to LEGACY_OFFCHAIN and\nstays there until it's actually linked on-chain (Escrow.contract_address\n/ Prediction.contract_address / Dispute.on_chain_dispute_id set) \u2014 see\nROADMAP.md 4.5's deferred \"add this together with the actual cutover\"\nnote; this is that column, added once there's a real indexer to drive\nit rather than speculatively ahead of one.",
        "enum": [
          "legacy_offchain",
          "processing",
          "decided",
          "finalized",
          "canceled"
        ],
        "title": "ChainStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OnChainBetAck": {
        "description": "Body for POST /predictions/{id}/bet/on-chain \u2014 the frontend\nreporting a tx hash it already got back from signing and sending\nNuancePredictionMarket.bet itself (see components/app/\ngenlayer-write-client.ts's betOnChain). Unlike a regular write ack,\nthis ALSO carries side/amount: the real stake already lives in the\ncontract's own storage, but this app still mirrors it into a\nPredictionPosition row immediately (same reasoning\nOnChainSubmissionAck's docstring gives for milestones) so the existing\n\"my positions\" UI keeps working without waiting on an indexer cycle\nthat doesn't sync individual bettors' stakes at all today.",
        "properties": {
          "amount": {
            "exclusiveMinimum": 0.0,
            "maximum": 1000000.0,
            "title": "Amount",
            "type": "integer"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "tx_hash": {
            "title": "Tx Hash",
            "type": "string"
          }
        },
        "required": [
          "tx_hash",
          "side",
          "amount"
        ],
        "title": "OnChainBetAck",
        "type": "object"
      },
...
      "PredictionBetCreate": {
        "properties": {
          "amount": {
            "exclusiveMinimum": 0.0,
            "maximum": 1000000.0,
            "title": "Amount",
            "type": "integer"
          },
          "side": {
            "title": "Side",
            "type": "string"
          }
        },
        "required": [
          "side",
          "amount"
        ],
        "title": "PredictionBetCreate",
        "type": "object"
      },
      "PredictionPositionRead": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "payout": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": 0.0,
            "title": "Payout"
          },
          "prediction_id": {
            "title": "Prediction Id",
            "type": "integer"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "status": {
            "default": "PENDING",
            "title": "Status",
            "type": "string"
          },
          "wallet_address": {
            "title": "Wallet Address",
            "type": "string"
          }
        },
        "required": [
          "id",
          "prediction_id",
          "wallet_address",
          "side",
          "amount",
          "created_at"
        ],
        "title": "PredictionPositionRead",
        "type": "object"
      },
      "PredictionRead": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "chain_status": {
            "$ref": "#/components/schemas/ChainStatus",
            "default": "legacy_offchain"
          },
          "contract_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Address"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "positions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PredictionPositionRead"
            },
            "title": "Positions",
            "type": "array"
          },
          "resolution_date": {
            "format": "date-time",
            "title": "Resolution Date",
            "type": "string"
          },
          "resolution_reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reasoning"
          },
          "resolution_source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Source Url"
          },
          "resolution_trigger_tx_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Trigger Tx Hash"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "status_key": {
            "title": "Status Key",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "volume": {
            "title": "Volume",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "category",
          "resolution_date",
          "volume",
          "status_key",
          "created_at"
        ],
        "title": "PredictionRead",
        "type": "object"
      },
...
      "WebhookCreate": {
        "properties": {
          "event_types": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Event Types",
            "type": "array"
          },
          "url": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "event_types"
        ],
        "title": "WebhookCreate",
        "type": "object"
      },
      "WebhookCreateResponse": {
        "description": "POST /webhooks's response \u2014 the only time `secret` is ever\nreturned; see models.core.Webhook's own docstring on why this one\n(unlike an ApiKey's secret) can't just be hashed: it has to stay\nusable server-side to keep signing every future delivery. Callers\nthat lose it must delete and re-register the webhook to get a new one\n(there's no \"rotate secret\" endpoint yet \u2014 out of scope for this\npass).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_types": {
            "items": {
              "type": "string"
            },
            "title": "Event Types",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Delivered At"
          },
          "last_delivery_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Delivery Status"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "url",
          "event_types",
          "is_active",
          "created_at",
          "secret"
        ],
        "title": "WebhookCreateResponse",
        "type": "object"
      },
      "WebhookRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_types": {
            "items": {
              "type": "string"
            },
            "title": "Event Types",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Delivered At"
          },
          "last_delivery_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Delivery Status"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "url",
          "event_types",
          "is_active",
          "created_at"
        ],
        "title": "WebhookRead",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend for Nuance \u2014 AI-validator consensus for claims traditional smart contracts can't settle.",
    "title": "Nuance API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/predictions": {
      "get": {
        "description": "Defaults to only `status_key == \"open\"` markets so unreviewed drafts\n(`\"pending_review\"`, see services/market_generator.py) never leak into\nthe public betting feed just because a caller forgot to filter.\n\n`status=None` or `status=\"all\"` (case-insensitive) returns every market\nregardless of status \u2014 e.g. for an internal review queue. Any other\nvalue filters to that exact status, matched case-insensitively since\nstatus_key casing isn't consistent across the codebase today (markets\nare created as lowercase \"open\"/\"pending_review\", but\nservices/prediction_oracle.py resolves them to uppercase \"RESOLVED\").",
        "operationId": "list_predictions_predictions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "open",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PredictionRead"
                  },
                  "title": "Response List Predictions Predictions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Predictions",
        "tags": [
          "predictions"
        ]
      }
    },
    "/predictions/{prediction_id}": {
      "get": {
        "operationId": "get_prediction_predictions__prediction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "prediction_id",
            "required": true,
            "schema": {
              "title": "Prediction Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Prediction",
        "tags": [
          "predictions"
        ]
      }
    },
    "/predictions/{prediction_id}/bet": {
      "post": {
        "operationId": "place_bet_predictions__prediction_id__bet_post",
        "parameters": [
          {
            "in": "path",
            "name": "prediction_id",
            "required": true,
            "schema": {
              "title": "Prediction Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionBetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Place Bet",
        "tags": [
          "predictions"
        ]
      }
    },
    "/predictions/{prediction_id}/bet/on-chain": {
      "post": {
        "description": "The on-chain counterpart to place_bet above \u2014 reached once\ncomponents/app/genlayer-write-client.ts's betOnChain has already\nsigned and sent a real NuancePredictionMarket.bet transaction directly\nto the chain. Doesn't verify the hash is real (see\nOnChainSubmissionAck's own docstring on why that's fine \u2014 nothing here\nis a trust boundary for the market's actual outcome); this only\nmirrors the stake into a PredictionPosition row so the existing \"my\npositions\" UI keeps working, since services/genlayer_indexer.py's\nview-sync doesn't track individual bettors' on-chain stakes today,\nonly the market's own state/outcome as a whole.",
        "operationId": "place_bet_on_chain_predictions__prediction_id__bet_on_chain_post",
        "parameters": [
          {
            "in": "path",
            "name": "prediction_id",
            "required": true,
            "schema": {
              "title": "Prediction Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnChainBetAck"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Place Bet On Chain",
        "tags": [
          "predictions"
        ]
      }
    },
    "/predictions/{prediction_id}/resolve": {
      "post": {
        "operationId": "resolve_prediction_predictions__prediction_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "prediction_id",
            "required": true,
            "schema": {
              "title": "Prediction Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Prediction",
        "tags": [
          "predictions"
        ]
      }
    },
...
    "/webhooks": {
      "get": {
        "operationId": "list_webhooks_webhooks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WebhookRead"
                  },
                  "title": "Response List Webhooks Webhooks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Webhooks",
        "tags": [
          "webhooks"
        ]
      },
      "post": {
        "operationId": "create_webhook_webhooks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Webhook",
        "tags": [
          "webhooks"
        ]
      }
    },
    "/webhooks/{webhook_id}": {
      "delete": {
        "operationId": "delete_webhook_webhooks__webhook_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "webhook_id",
            "required": true,
            "schema": {
              "title": "Webhook Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Webhook",
        "tags": [
          "webhooks"
        ]
      }
    }
  }
}